# backend/services/job-service/swagger.yaml
openapi: 3.0.3
info:
  title: LaborGuard Job Service API
  description: Job Management Service for LaborGuard
  version: 1.0.0
servers:
  - url: http://localhost:5003
...
paths:
  /api/jobs:
    get:
      tags: [Jobs]
      summary: Get all jobs
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [open, closed, filled, pending_review]
        - in: query
          name: employerId
          schema:
            type: string
        - in: query
          name: jobType
          schema:
            type: string
            enum: [full-time, part-time, contract, temporary, freelance]
      responses:
        "200":
          description: List of jobs
    post:
      tags: [Jobs]
      summary: Create a new job posting
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, wage, location, jobType]
              properties:
                title:
                  type: string
                  maxLength: 100
                description:
                  type: string
                  maxLength: 1000
                wage:
                  type: object
                  required: [amount, frequency]
                  properties:
                    amount:
                      type: number
                      minimum: 0
                    currency:
                      type: string
                      default: USD
                    frequency:
                      type: string
                      enum: [hourly, daily, weekly, monthly, yearly, per-task]
                compliesWithMinimumWage:
                  type: boolean
                location:
                  type: object
                  required: [address, city]
                  properties:
                    address:
                      type: string
                    city:
                      type: string
                    state:
                      type: string
                    country:
                      type: string
                      default: USA
                jobType:
                  type: string
                  enum: [full-time, part-time, contract, temporary, freelance]
      responses:
        "201":
          description: Job created
        "401":
          description: Unauthorized
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT